LPS-162388 Immediately return empty link if user should not have access
If a notification feed entry's applicable is false it will show marked
as read, have no link and no actions. It was meant for ephemeral
notifications associated to workflow tasks that were withdrawn.

This is too stringent for notifications meant to have no link to another
view but still have the option to mark as read and delete.

For these, applicable should be true and link should be empty. This way
the notification is marked unread, has actions and is given generated
link that doesn't go anywhere but marks the notification read.